+++ projects/np-ui-lib/src/lib/np-paginator/np-paginator.component.ts
@@ -1,0 +1,96 @@
+import { Component, Input, Output, EventEmitter, OnInit, ViewEncapsulation, ChangeDetectionStrategy, OnChanges, SimpleChanges } from '@angular/core';
+
+@Component({
+  selector: 'np-paginator',
+  templateUrl: './np-paginator.component.html',
+  styleUrls: ['./np-paginator.component.css'],
+  encapsulation: ViewEncapsulation.None,
+  changeDetection: ChangeDetectionStrategy.Default
+})
+export class NpPaginatorComponent implements OnInit, OnChanges {
+  static controlCount = 1;
+
+  @Input() pageSize: number;
+  @Input() totalItems: number;
+  @Input() currentPage: number;
+  @Input() showPageSize = true;
+  @Input() showTotal = true;
+  @Input() styleClass: string;
+  @Input() inputId = `np-paginator-box_${NpPaginatorComponent.controlCount++}`;
+
+  @Output() onPageChange: EventEmitter<any> = new EventEmitter();
+
+  totalPages: number;
+  inited = false;
+
+  constructor() { }
+
+  ngOnChanges(changes: SimpleChanges): void {
+    if (changes && changes.totalItems) {
+      this.totalPages = this.calculateTotalPages();
+      if (this.currentPage > this.totalPages) {
+        this._applyChanges();
+      }
+      return;
+    }
+    if (changes && (changes.currentPage || changes.pageSize)) {
+      this._applyChanges();
+    }
+  }
+
+  ngOnInit(): void {
+    if (!this.pageSize) {
+      this.pageSize = 10;
+    }
+    if (!this.totalItems) {
+      this.totalItems = 0;
+    }
+    if (!this.currentPage) {
+      this.currentPage = 1;
+    }
+    this.inited = true;
+    this._applyChanges();
+  }
+
+  _onCurrentPageChange(event) {
+    this._loadPage(Number(event.target.value));
+  }
+
+  _onItemPerPageChange(event) {
+    this.pageSize = Number(event.target.value);
+    this._loadPage(1);
+  }
+
+  _loadPage(page: number) {
+    this.currentPage = page;
+    this._applyChanges();
+  }
+
+  _applyChanges() {
+    if (this.inited) {
+      this.totalPages = this.calculateTotalPages();
+    }
+    if (this.currentPage > 0 && this.pageSize > 0 && this.totalPages > 0) {
+      if (this.currentPage < 1) {
+        this.currentPage = 1;
+      }
+      if (this.currentPage > this.totalPages) {
+        this.currentPage = this.totalPages;
+      }
+      this.onPageChange.emit({ currentPage: this.currentPage, pageSize: this.pageSize });
+    }
+  }
+
+  loadPage(page: number) {
+    this._loadPage(page);
+  }
+
+  refresh() {
+    this.onPageChange.emit({ currentPage: this.currentPage, pageSize: this.pageSize });
+  }
+
+  private calculateTotalPages(): number {
+    const result = this.pageSize < 1 ? 1 : Math.ceil(this.totalItems / this.pageSize);
+    return Math.max(result || 0, 1);
+  }
+}