+++ src/app/np-paginator-demo/np-paginator-demo-routing.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { NpPaginatorDemoComponent } from './np-paginator-demo.component';
+
+const routes: Routes = [{
+  path: '',
+  component: NpPaginatorDemoComponent
+}];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class NpPaginatorDemoRoutingModule { }